requirements.txt
feedparser
newsapi-python
pandas
python-dotenv
